+++ src/modules/modem/analog/ModemUSB.cpp
@@ -49,7 +49,7 @@ void ModemUSB::demodulate(ModemKit *kit, ModemIQData *input, AudioThreadInput *a
     }
     
     liquid_float_complex x, y;
-    for (int i = 0; i < bufSize; i++) { // Reject lower band
+    for (size_t i = 0; i < bufSize; i++) { // Reject lower band
         nco_crcf_step(ssbShift);
         nco_crcf_mix_down(ssbShift, input->data[i], &x);
         iirfilt_crcf_execute(ssbFilt, x, &y);
