+++ src/modules/modem/digital/ModemST.cpp
@@ -20,7 +20,7 @@ void ModemST::demodulate(ModemKit *kit, ModemIQData *input, AudioThreadInput *au
     ModemKitDigital *dkit = (ModemKitDigital *)kit;
     digitalStart(dkit, demodST, input);
 
-    for (int i = 0, bufSize = input->data.size(); i < bufSize; i++) {
+    for (size_t i = 0, bufSize = input->data.size(); i < bufSize; i++) {
         modem_demodulate(demodST, input->data[i], &demodOutputDataDigital[i]);
     }
     updateDemodulatorLock(demodST, 0.005f);
